[WIP] Add abstraction layer for cluster
PRosettaC was written to explicitely rely on PBS for job scheduling.
This prevented people from running it with other schedulers.

This patch overcomes this by introducing a layer of abstaction
between the main algorithm and the code actually talking to the
scheduler. This way, it's easy to implement support for different
schedulers and let PRosettaC select the correct one at runtime.

With some default values we make sure that, from a user's point of
view, the code behaves exactly the same as before. So when you run it
with the same configuration on PBS before and after this patch, the
behavior does not change. For this, parsing of the config file has to
be changed slightly to easily spot unset optional parameters.

We also give the user some control over what's included in the
generated job files sent to the scheduler (see Readme). This includes
parameters RosettaDockMemory and ProtacModelMemory, and also a new
environment variable SCHEDULER_PARAMS. All this is optional.

The PBS-specific code, formerly in PBS/cluster_pbs.py, is now located
in cluster/PBS/PBS.py.